requirements.txt
python-dotenv>=1.0
openai>=1.40
numpy>=1.26
tqdm>=4.66
pdfminer.six>=20231228
trafilatura>=2.0
